+++ src/django_idom/components.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import json
 import os
 from inspect import iscoroutinefunction
-from typing import Any, Callable, Dict, Iterable
+from typing import Any, Callable, Dict, Protocol, Sequence
 
 from channels.db import database_sync_to_async
 from django.contrib.staticfiles.finders import find
@@ -11,48 +11,45 @@ from django.http import HttpRequest
 from django.urls import reverse
 from django.views import View
 from idom import component, hooks, html, utils
-from idom.types import VdomDict
+from idom.types import ComponentType, VdomDict
 
 from django_idom.config import IDOM_CACHE, IDOM_VIEW_COMPONENT_IFRAMES
 from django_idom.types import ViewComponentIframe
 from django_idom.utils import _generate_obj_name
 
 
-# TODO: Might want to intercept href clicks and form submit events.
-# Form events will probably be accomplished through the upcoming DjangoForm.
+class _ViewComponentConstructor(Protocol):
+    def __call__(
+        self, request: HttpRequest | None, *args: Any, **kwargs: Any
+    ) -> ComponentType:
+        ...
+
+
 @component
-def view_to_component(
+def _view_to_component(
     view: Callable | View,
-    compatibility: bool = False,
-    transforms: Iterable[Callable[[VdomDict], Any]] = (),
-    strict_parsing: bool = True,
-    request: HttpRequest | None = None,
-    args: Iterable = (),
-    kwargs: Dict | None = None,
-) -> VdomDict | None:
-    """Converts a Django view to an IDOM component.
-
-    Args:
-        view: The view function or class to convert.
-
-    Keyword Args:
-        compatibility: If True, the component will be rendered in an iframe.
-            When using compatibility mode `tranforms`, `strict_parsing`, and `request`
-            arguments will be ignored.
-        transforms: A list of functions that transforms the newly generated VDOM.
-            The functions will be called on each VDOM node.
-        strict_parsing: If True, an exception will be generated if the HTML does not
-            perfectly adhere to HTML5.
-        request: Request object to provide to the view.
-        args: The positional arguments to pass to the view.
-        kwargs: The keyword arguments to pass to the view.
-    """
+    compatibility: bool,
+    transforms: Sequence[Callable[[VdomDict], Any]],
+    strict_parsing: bool,
+    request: HttpRequest | None,
+    args: Sequence | None,
+    kwargs: Dict | None,
+):
+    converted_view, set_converted_view = hooks.use_state(None)
+    args = args or ()
     kwargs = kwargs or {}
-    rendered_view, set_rendered_view = hooks.use_state(None)
     request_obj = request
-    if not request:
+    if not request_obj:
         request_obj = HttpRequest()
         request_obj.method = "GET"
+    if compatibility:
+        dotted_path = f"{view.__module__}.{view.__name__}"  # type: ignore[union-attr]
+        dotted_path = dotted_path.replace("<", "").replace(">", "")
+        IDOM_VIEW_COMPONENT_IFRAMES[dotted_path] = ViewComponentIframe(
+            view, args, kwargs
+        )
+    else:
+        dotted_path = None
 
     # Render the view render within a hook
     @hooks.use_effect(
@@ -61,18 +58,11 @@ def view_to_component(
             json.dumps([args, kwargs], default=lambda x: _generate_obj_name(x)),
         ]
     )
-    async def async_renderer():
+    async def async_render():
         """Render the view in an async hook to avoid blocking the main thread."""
         # Render Check 1: Compatibility mode
         if compatibility:
-            dotted_path = f"{view.__module__}.{view.__name__}"
-            dotted_path = dotted_path.replace("<", "").replace(">", "")
-            IDOM_VIEW_COMPONENT_IFRAMES[dotted_path] = ViewComponentIframe(
-                view, args, kwargs
-            )
-
-            # Signal that the view has been rendered
-            set_rendered_view(
+            set_converted_view(
                 html.iframe(
                     {
                         "src": reverse("idom:view_to_component", args=[dotted_path]),
@@ -84,43 +74,96 @@ def view_to_component(
 
         # Render Check 2: Async function view
         elif iscoroutinefunction(view):
-            render = await view(request_obj, *args, **kwargs)
+            view_html = await view(request_obj, *args, **kwargs)
 
         # Render Check 3: Async class view
         elif getattr(view, "view_is_async", False):
             view_or_template_view = await view.as_view()(request_obj, *args, **kwargs)
             if getattr(view_or_template_view, "render", None):  # TemplateView
-                render = await view_or_template_view.render()
+                view_html = await view_or_template_view.render()
             else:  # View
-                render = view_or_template_view
+                view_html = view_or_template_view
 
         # Render Check 4: Sync class view
         elif getattr(view, "as_view", None):
             async_cbv = database_sync_to_async(view.as_view())
             view_or_template_view = await async_cbv(request_obj, *args, **kwargs)
             if getattr(view_or_template_view, "render", None):  # TemplateView
-                render = await database_sync_to_async(view_or_template_view.render)()
+                view_html = await database_sync_to_async(view_or_template_view.render)()
             else:  # View
-                render = view_or_template_view
+                view_html = view_or_template_view
 
         # Render Check 5: Sync function view
         else:
-            render = await database_sync_to_async(view)(request_obj, *args, **kwargs)
+            view_html = await database_sync_to_async(view)(request_obj, *args, **kwargs)
 
         # Signal that the view has been rendered
-        set_rendered_view(
+        set_converted_view(
             utils.html_to_vdom(
-                render.content.decode("utf-8").strip(),
+                view_html.content.decode("utf-8").strip(),
                 *transforms,
                 strict=strict_parsing,
             )
         )
 
-    # Return the view if it's been rendered via the `async_renderer` hook
-    return rendered_view
+    # Return the view if it's been rendered via the `async_render` hook
+    return converted_view
+
+
+# TODO: Might want to intercept href clicks and form submit events.
+# Form events will probably be accomplished through the upcoming DjangoForm.
+def view_to_component(
+    view: Callable | View = None,  # type: ignore[assignment]
+    compatibility: bool = False,
+    transforms: Sequence[Callable[[VdomDict], Any]] = (),
+    strict_parsing: bool = True,
+) -> _ViewComponentConstructor:
+    """Converts a Django view to an IDOM component.
+
+    Keyword Args:
+        view: The view function or class to convert.
+        compatibility: If True, the component will be rendered in an iframe.
+            When using compatibility mode `tranforms`, `strict_parsing`, `request`,
+            `args, and `kwargs` arguments will be ignored.
+        transforms: A list of functions that transforms the newly generated VDOM.
+            The functions will be called on each VDOM node.
+        strict_parsing: If True, an exception will be generated if the HTML does not
+            perfectly adhere to HTML5.
+
+    Returns:
+        Callable: A function that takes `request: HttpRequest | None, *args: Any, **kwargs: Any`
+            and returns an IDOM component.
+    """
+
+    def decorator(view: Callable | View):
+        if not view:
+            raise ValueError("A view must be provided to `view_to_component`")
+
+        def wrapper(
+            request: HttpRequest | None = None,
+            *args: Any,
+            **kwargs: Any,
+        ):
+            return _view_to_component(
+                view=view,
+                compatibility=compatibility,
+                transforms=transforms,
+                strict_parsing=strict_parsing,
+                request=request,
+                args=args,
+                kwargs=kwargs,
+            )
+
+        return wrapper
+
+    return decorator(view) if view else decorator
 
 
 @component
+def _django_css(static_path: str):
+    return html.style(_cached_static_contents(static_path))
+
+
 def django_css(static_path: str):
     """Fetches a CSS static file for use within IDOM. This allows for deferred CSS loading.
 
@@ -128,10 +171,15 @@ def django_css(static_path: str):
         static_path: The path to the static file. This path is identical to what you would
         use on a `static` template tag.
     """
-    return html.style(_cached_static_contents(static_path))
+
+    return _django_css(static_path=static_path)
 
 
 @component
+def _django_js(static_path: str):
+    return html.script(_cached_static_contents(static_path))
+
+
 def django_js(static_path: str):
     """Fetches a JS static file for use within IDOM. This allows for deferred JS loading.
 
@@ -139,7 +187,8 @@ def django_js(static_path: str):
         static_path: The path to the static file. This path is identical to what you would
         use on a `static` template tag.
     """
-    return html.script(_cached_static_contents(static_path))
+
+    return _django_js(static_path=static_path)
 
 
 def _cached_static_contents(static_path: str):