+++ tests/test_app/tests/test_components.py
@@ -182,3 +182,11 @@ class TestIdomCapabilities(ChannelsLiveServerTestCase):
         ).locator(
             "#ViewToComponentTemplateViewClassCompatibility[data-success=true]"
         ).wait_for()
+
+    def test_view_to_component_decorator(self):
+        self.page.locator("#view_to_component_decorator[data-success=true]").wait_for()
+
+    def test_view_to_component_decorator_args(self):
+        self.page.locator(
+            "#view_to_component_decorator_args[data-success=true]"
+        ).wait_for()